'Flow-based Botnet Detection System' with eBPF
The Flow-based Botnet Detection System was connected to the eBPF program: it receives the captured traffic from the eBPF program and passes it to Random Forest Classifier for the classification. The class 'TaskQueue' (in the file 'utilities/task_queue.py') allows to enqueue in a Queue the pieces of captured traffic to analyze one at a time.

Signed-off-by: Alessio Izzillo <alessioizzillo@gmail.com>