Modified AntiBotnet.py and fixed Flow Based Detection
Problems:
 - the validity of the arguments passed while launching AntiBotnet.py, is not
   properly checked;
 - the Flow Features Extractor is slow to extract features from a large
   dataset.

A proper validation of the AntiBotnet.py arguments was added. The performance
of the Flow Features Extractor was improved by using lists while looping for
computing the features and the pandas Dataframe at the end of the computation.
In addition, the Random Forest Classifier was "separated" better from the
Flow-based detection method in order to improve the modularity to allow the
usage by the future implementation of Graph-based Detection method.

Signed-off-by: Alessio Izzillo <alessioizzillo@gmail.com>