A first version of the Graph-based Botnet Detection System
A first version of the Graph-based Detection method was implemented by
developing a module for creating graph structures ("create_graph.py") and a
module for extracting the features from the created graphs
("graph_features_extractor.py"). In addition, a P2P system (using the Gnutella
protocol) was added to allow the hosts belonging to it to exchange among them
the capture traffic in order to perform the predicition not just on the local
traffic but on the union of the local traffic and the received traffics of the
other hosts of the P2P network.

Signed-off-by: Alessio Izzillo <alessioizzillo@gmail.com>