Graph-based Oracle
The "GraphBasedOracle.py" file was added for analyzing (with the Graph-based
method) the "test_traffic.csv" produced by running the tool in "TEST MODE". In
this way, it's possible to make a comparison (in terms of performance and
accuracy) between an analysis of the complete traffic (performed by
"GraphBasedOracle.py") and an analysis using Incremental Learning (performed by
the tool).

Signed-off-by: Alessio Izzillo <alessioizzillo@gmail.com>